move controls to different component
